fix: add revert missing blocks for persisting blocks to prevent storing the missing block type definition
