+++ src/FormBuilder/resources/js/form-builder/src/common/revertMissingBlocks.ts
@@ -1,0 +1,22 @@
+import {BlockInstance, getBlockType, getUnregisteredTypeHandlerName} from '@wordpress/blocks';
+
+/**
+ * Replaces any missing blocks with the original block type.
+ *
+ * This is used to preserve the original block type definition when persisting the blocks to the database.
+ *
+ * @unreleased
+ */
+export default function revertMissingBlocks(blocks: BlockInstance[]) {
+    blocks.forEach((sectionBlock, sectionBlockIndex) => {
+        sectionBlock.innerBlocks?.forEach((innerBlock, innerBlockIndex) => {
+            const blockType = getBlockType(innerBlock.name);
+
+            if (blockType.name === getUnregisteredTypeHandlerName() && innerBlock.attributes?.originalContent) {
+                blocks[sectionBlockIndex].innerBlocks[innerBlockIndex] = JSON.parse(
+                    innerBlock.attributes.originalContent
+                );
+            }
+        });
+    });
+}
